pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adversarial-curriculum-offline-rl-with-uncertainty-weighted-conservative-learning"
version = "0.1.0"
description = "Offline RL with adversarial curriculum and uncertainty-weighted conservative learning"
authors = [{name = "Alireza Shojaei"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "gymnasium>=0.29.0",
    "d4rl>=1.1",
    "stable-baselines3>=2.0.0",
    "tianshou>=0.5.0",
    "pyyaml>=6.0",
    "mlflow>=2.8.0",
    "tensorboard>=2.14.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools]
packages = ["adversarial_curriculum_offline_rl_with_uncertainty_weighted_conservative_learning"]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
